refactor all of RatingsAndReviews
